pyproject.toml
[project]
name = "cube"
version = "1.0.0"
description = "三阶魔方速解（基于 CFOP 方法）"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "rubiktwophase>=1.1.1",
    "torch>=2.9.1",
    "ultralytics>=8.3.234",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=9.0.1"]
